requirements.txt
asyncpg==0.29.0
celery==5.4.0
fastapi==0.109.2
fastapi-cli==0.0.4
fiona==1.9.6
flower==1.2.0
GeoAlchemy2==0.15.1
geojson==3.1.0
geopandas==0.14.4
gpxpy==1.6.2
momepy==0.7.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.3
osmnx==1.9.3
pandas==2.2.2
psycopg2==2.9.9
pyarrow==16.1.0
pydantic==2.7.3
python-dotenv==1.0.1
pydantic_core==2.18.4
redis==5.0.6
referencing==0.35.1
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
SRTM.py==0.3.7
sse-starlette==2.1.2
SQLAlchemy==2.0.30
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uvicorn==0.27.1
requests==2.31.0
tqdm==4.66.4
python-multipart==0.0.9